Fixed bug in differential deviation mode of kinematics generation.

+++ lib/yawff.c
@@ -943,10 +943,10 @@ int update_wing_kine(
       u_clamped = u_clamped > -DIFF_DEV_MAX_U ? u : -DIFF_DEV_MAX_U;
 
       // Dummy values need to set this to actual kinematics
-      vals[STROKE_0_ID] = (str_amp/asinf(str_k))*asinf(str_k*cosf(2.0*M_PI*t/T));
-      vals[STROKE_1_ID] = (str_amp/asinf(str_k))*asinf(str_k*cosf(2.0*M_PI*t/T));
-      vals[ROTATION_0_ID] = (rot_amp/tanhf(rot_k))*tanhf(rot_k*sinf(2.0*M_PI*t/T));
-      vals[ROTATION_1_ID] = (rot_amp/tanhf(rot_k))*tanhf(rot_k*sinf(2.0*M_PI*t/T));
+      vals[STROKE_0_ID] = (str_amp/asinf(str_k))*asinf(str_k*((float)cos(2.0*M_PI*t/(double)T)));
+      vals[STROKE_1_ID] = (str_amp/asinf(str_k))*asinf(str_k*((float)cos(2.0*M_PI*t/(double)T)));
+      vals[ROTATION_0_ID] = (rot_amp/tanhf(rot_k))*tanhf(rot_k*((float)sin(2.0*M_PI*t/(double)T)));
+      vals[ROTATION_1_ID] = (rot_amp/tanhf(rot_k))*tanhf(rot_k*((float)sin(2.0*M_PI*t/(double)T)));
       vals[DEVIATION_0_ID] = u_clamped;
       vals[DEVIATION_1_ID] = -u_clamped; 
      break;
